feat(storefront): add type scale pattern page | sas
