+++ packages/storefront/src/main.ts
@@ -13,7 +13,6 @@ import PartialDocs from '@/components/PartialDocs.vue';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import InstantSearch from 'vue-instantsearch';
-import TypeScale from '@/components/TypeScale.vue';
 
 Vue.use(InstantSearch);
 
@@ -45,7 +44,6 @@ Vue.use({
     vue.component('OpenBugTemplateInStackBlitz', OpenBugTemplateInStackBlitz);
     vue.component('TableOfContents', TableOfContents);
     vue.component('PartialDocs', PartialDocs);
-    vue.component('TypeScale', TypeScale);
   },
 });
